+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# main
+  * Fix behavior when a window event handler destroys its window.
+
 # v0.12.1 - 2022-04-15
   * Change `Container` for `image` crate support to accept any `Deref<Target = [u8]` container.